Update READme
The method is called scoring with an array passed as an argument. Elements of the array are of the class User.

User class has a method update_score.

The task is to iterate through each of the elements in the array using each and call the method update_score on every element.